commands refactor
