+++ src/cli/mod.rs
@@ -2,8 +2,10 @@ use clap::Parser;
 use miden_client::{client::Client, config::ClientConfig};
 
 mod account;
+mod info;
 mod input_notes;
-mod sync_state;
+mod sync;
+mod tags;
 mod transactions;
 
 /// Root CLI struct
@@ -26,8 +28,12 @@ pub enum Command {
     Account(account::AccountCmd),
     #[clap(subcommand)]
     InputNotes(input_notes::InputNotes),
+    /// Sync this client with the latest state of the Miden network.
+    Sync,
+    /// View a summary of the current client state
+    Info,
     #[clap(subcommand)]
-    SyncState(sync_state::SyncStateCmd),
+    Tags(tags::TagsCmd),
     #[clap(subcommand)]
     Transaction(transactions::Transaction),
     #[cfg(feature = "mock")]
@@ -49,8 +55,10 @@ impl Cli {
         // execute cli command
         match &self.action {
             Command::Account(account) => account.execute(client),
+            Command::Info => info::print_client_info(&client),
             Command::InputNotes(notes) => notes.execute(client),
-            Command::SyncState(tags) => tags.execute(client).await,
+            Command::Sync => sync::sync_state(client).await,
+            Command::Tags(tags) => tags.execute(client).await,
             Command::Transaction(transaction) => transaction.execute(client).await,
             #[cfg(feature = "mock")]
             Command::MockData { transaction } => {
